[#43] Initial styling
