--- src/components/MapReference/InnerComponents/BlackenedBackground.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function BlackenedBackground(){
-    return <div className={"blackened-background"}></div>
-}
