+++ src/components/MapReference/InnerComponents/BlackenedBackground/BlackenedBackground.css
@@ -1,0 +1,6 @@
+.blackened-background {
+    width: 100vw;
+    height: 100vh;
+    position: absolute;
+    background-color: rgba(0, 0, 0, 0.5);
+}
