+++ src/components/MapReference/InnerComponents/BlackenedBackground/BlackenedBackground.jsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import './BlackenedBackground.css'
+
+export default function BlackenedBackground(){
+    return <div className={"blackened-background"} />
+}
