+++ src/components/MapReference/InnerComponents/ContentContainer/ContentContainer.jsx
@@ -1,4 +1,5 @@
 import React, { useState } from 'react';
+import './ContentContainer.css'
 
 // Components
 import MapInformation from "./InnerComponents/MapContent/MapInformation";
@@ -11,23 +12,16 @@ export default function ContentContainer(props){
         setCurrentMap(event.target.id);
     }
 
-    const mapList = props.mapData.map((location) => {
-        return(
-        <MapList
-            mapId={location.id}
-            mapName={location.name}
-            mapSize={location.size}
-            handleCurrentMap={handleCurrentMap}
-        />)
-    });
-
     const currentMapInformation = currentMap !== 'none'
         ? props.map.filter((location) => { return location.id === currentMap })
         : 'none';
 
     return (
         <div className={"map-content-container"}>
-            { mapList }
+            <MapList
+                mapData={props.mapData}
+                handleCurrentMap={handleCurrentMap}
+            />
             <MapInformation mapInfo={currentMapInformation} />
         </div>
     )
