+++ src/components/MapReference/InnerComponents/ContentContainer/InnerComponents/MapList/MapListItem.jsx
@@ -1,0 +1,5 @@
+export default function MapListItem(props){
+    const itemClass = props.activeMap ? "map-list-item-active" : "map-list-item-inactive";
+
+    return <li id={props.mapId} className={itemClass} onClick={props.handleCurrentMap}>{props.mapName}</li>
+}
