--- src/components/MapReference/InnerComponents/InnerComponents/MapList/MapList.jsx
@@ -1,5 +1,0 @@
-import React from 'react';
-
-export default function MapList(props){
-    return null
-}
