+++ src/components/MapReference/MapReference.css
@@ -1,0 +1,6 @@
+.map-reference-main-container {
+    position: absolute;
+    z-index: 1;
+    width: 100vw;
+    height: 100vh;
+}
