+++ src/components/MapReference/MapReference.jsx
@@ -1,17 +1,18 @@
 import React from 'react';
+import './MapReference.css';
 
 // Data
 import mapData from '../../utils/data/mapData.json';
 
 // Components
-import BlackenedBackground from "./InnerComponents/BlackenedBackground";
-import ContentContainer from "./InnerComponents/ContentContainer";
+import BlackenedBackground from "./InnerComponents/BlackenedBackground/BlackenedBackground";
+import ContentContainer from "./InnerComponents/ContentContainer/ContentContainer";
 
 export default function MapReference(){
     return (
         <div className={"map-reference-main-container"}>
-            <BlackenedBackground />
             <ContentContainer mapData={mapData.maps} />
+            <BlackenedBackground />
         </div>
     )
 }
